fix: BibbiSegmentedControl AutoLayout 수정 및 StudioButton 제거 - BibbiSegmentedControl+Ext 파일 추가 및 ControlProperty, Binder 파일 내부에 추가
