+++ 14th-team5-iOS/Core/Sources/Bibbi/BBCommons/BBSegmentedControl/BibbiSegmentedControl.swift
@@ -28,7 +28,9 @@ public final class BibbiSegmentedControl: UIView {
     // MARK: - Views
     internal let survivalButton = UIButton(type: .system)
     internal let missionButton  = UIButton(type: .system)
-    internal let studioButton   = UIButton(type: .system)
+    
+    //TODO: 다음 달에 추가해서 사용해주세요!
+    //internal let studioButton   = UIButton(type: .system)
 
     // MARK: - Private
     private let disposeBag = DisposeBag()
@@ -38,7 +40,7 @@ public final class BibbiSegmentedControl: UIView {
         super.init(frame: frame)
         setupUI()
         setupAttributes()
-        setupLayout()
+        setupAutoLayout()
         bind()
         applySelection(selected.value)
     }
@@ -47,7 +49,7 @@ public final class BibbiSegmentedControl: UIView {
         super.init(coder: coder)
         setupUI()
         setupAttributes()
-        setupLayout()
+        setupAutoLayout()
         bind()
         applySelection(selected.value)
     }
@@ -56,7 +58,7 @@ public final class BibbiSegmentedControl: UIView {
 // MARK: - Setup
 private extension BibbiSegmentedControl {
     func setupUI() {
-        addSubviews(survivalButton, missionButton, studioButton)
+        addSubviews(survivalButton, missionButton)
     }
 
     func setupAttributes() {
@@ -84,10 +86,8 @@ private extension BibbiSegmentedControl {
             var c = btn.configuration
 
             if btn.isSelected {
-                if btn != self.studioButton {               // ← 스튜디오는 배경색/전경색 안 건드림
-                    btn.backgroundColor = .gray100
-                    c?.baseForegroundColor = .bibbiBlack
-                }
+                btn.backgroundColor = .gray100
+                c?.baseForegroundColor = .bibbiBlack
             } else {
                 btn.backgroundColor = .clear
                 c?.baseForegroundColor = .gray500
@@ -108,55 +108,58 @@ private extension BibbiSegmentedControl {
             $0.configurationUpdateHandler = updateHandler
         }
 
-        studioButton.do {
-            $0.configuration = .plain()
-            $0.configurationUpdateHandler = { btn in
-                var config = btn.configuration ?? .plain()
-
-                // 상태별 배경 이미지
-                var bg = UIBackgroundConfiguration.clear()
-                bg.image = btn.isSelected
-                    ? DesignSystemAsset.studioSelected.image
-                    : DesignSystemAsset.studioUnselected.image
-                bg.imageContentMode = .scaleAspectFill
-
-                config.background = bg
-                config.attributedTitle = nil
-                config.baseForegroundColor = .clear
-                config.contentInsets = .init(top: 0, leading: 0, bottom: 0, trailing: 0)
-
-                btn.configuration = config
-                btn.layer.cornerRadius = 20
-                btn.clipsToBounds = true
-            }
-        }
+        //TODO: 해당 부분도 다음달에 사용하기 때문에 주석처리했습니다.
+//        studioButton.do {
+//            $0.configuration = .plain()
+//            $0.configurationUpdateHandler = { btn in
+//                var config = btn.configuration ?? .plain()
+//
+//                // 상태별 배경 이미지
+//                var bg = UIBackgroundConfiguration.clear()
+//                bg.image = btn.isSelected
+//                    ? DesignSystemAsset.studioSelected.image
+//                    : DesignSystemAsset.studioUnselected.image
+//                bg.imageContentMode = .scaleAspectFill
+//
+//                config.background = bg
+//                config.attributedTitle = nil
+//                config.baseForegroundColor = .clear
+//                config.contentInsets = .init(top: 0, leading: 0, bottom: 0, trailing: 0)
+//
+//                btn.configuration = config
+//                btn.layer.cornerRadius = 20
+//                btn.clipsToBounds = true
+//            }
+//        }
 
     }
 
-    func setupLayout() {
+    func setupAutoLayout() {
         survivalButton.snp.makeConstraints {
-            $0.top.bottom.left.equalToSuperview()
+            $0.top.left.equalToSuperview()
             $0.width.equalTo(70)
+            $0.height.equalTo(40)
         }
 
         missionButton.snp.makeConstraints {
-            $0.top.bottom.equalToSuperview()
-            $0.leading.equalTo(survivalButton.snp.trailing)
+            $0.top.right.equalToSuperview()
+            $0.height.equalTo(40)
             $0.width.equalTo(70)
         }
 
-        studioButton.snp.makeConstraints {
-            $0.top.bottom.trailing.equalToSuperview()
-            $0.leading.equalTo(missionButton.snp.trailing)
-            $0.width.equalTo(94)
-        }
+        //TODO: 마찬가지
+//        studioButton.snp.makeConstraints {
+//            $0.top.bottom.trailing.equalToSuperview()
+//            $0.leading.equalTo(missionButton.snp.trailing)
+//            $0.width.equalTo(94)
+//        }
     }
 
     func bind() {
         let taps = Observable.merge(
-            survivalButton.rx.tap.map { SelectedSegment.survival },
-            missionButton.rx.tap.map { SelectedSegment.mission },
-            studioButton.rx.tap.map { SelectedSegment.studio }
+            survivalButton.rx.tap.map { SelectedSegment.survival }.debug("생존신고 터치 이벤트"),
+            missionButton.rx.tap.map { SelectedSegment.mission }.debug("미션 터치 이벤트")
+//            studioButton.rx.tap.map { SelectedSegment.studio }
         )
 
         // 탭 → 상태 갱신
@@ -183,50 +186,6 @@ private extension BibbiSegmentedControl {
     func applySelection(_ segment: SelectedSegment) {
         survivalButton.isSelected = (segment == .survival)
         missionButton.isSelected  = (segment == .mission)
-        studioButton.isSelected   = (segment == .studio)
-    }
-}
-
-// MARK: - Reactive Extensions
-public extension Reactive where Base: BibbiSegmentedControl {
-
-    /// SelectedSegment ↔︎ BibbiFeedType 양방향 바인딩
-    var selectedFeedType: ControlProperty<BibbiFeedType> {
-        let values = base.selected
-            .map { seg -> BibbiFeedType in
-                switch seg {
-                case .survival: return .survival
-                case .mission:  return .mission
-                case .studio:   return .studio
-                }
-            }
-
-        let setter = Binder<BibbiFeedType>(base) { control, feed in
-            switch feed {
-            case .survival: control.selected.accept(.survival)
-            case .mission:  control.selected.accept(.mission)
-            case .studio:   control.selected.accept(.studio)
-            }
-        }
-
-        return ControlProperty(values: values, valueSink: setter.asObserver())
-    }
-
-    /// 외부에서 programmatic 선택
-    var setSelectedType: Binder<BibbiFeedType> {
-        Binder(base) { control, type in
-            switch type {
-            case .survival: control.selected.accept(.survival)
-            case .mission:  control.selected.accept(.mission)
-            case .studio:   control.selected.accept(.studio)
-            }
-        }
-    }
-
-    /// 미션 탭 뱃지 토글
-    fileprivate var missionBadgeVisible: Binder<Bool> {
-        Binder(base) { control, visible in
-            control.missionButton.configuration?.image = visible ? DesignSystemAsset.mission.image : nil
-        }
+//        studioButton.isSelected   = (segment == .studio)
     }
 }
